feat: add keys utility to handle KeyboardEvent events
Fine tuning formatter exclusion and splitting library vs testing code
